new scripts for different spws

+++ C_Aarray/h2co_bigr.py
@@ -1,15 +1,16 @@
 """
-May 15, 2014: creation
+May 20, 2014: creation
 """
 # Split the H2CO window and its corresponding continuum out
 # (in order to avoid any risk of overwriting anything in the pipeline data)
 
-outputvis = 'h2co_35to80kms'
+outputvis = 'h2co_widr'
 if not os.path.exists(outputvis):
     vis = '../13A-064.sb28612538.eb29114303.56766.55576449074.ms'
     split(vis=vis, outputvis=outputvis, datacolumn='corrected',
-          spw='17:449~542', field='', width=1)
+          spw='17:349~642', field='', width=4)
 vis = outputvis
+
 
 phasecaltable = '../ch3oh/ch3oh_selfcal_phase09'
 ampcaltable = '../ch3oh/ch3oh_selfcal_ampphase'
@@ -22,7 +23,7 @@ flagdata(vis=vis, mode='clip', clipzeros=True)
 flagmanager(vis=vis, mode='save', versionname='cleanflags', comment='Flagged no antennae, zeros, and NOTHING from applycal.')
 
 clearcal(vis=vis)
-pfx = 'h2co_line_35to80kms.nocal'
+pfx = 'h2co_line_widr.nocal'
 os.system('rm -rf {0}.*'.format(pfx))
 clean(vis=vis, spw='0', imagename=pfx,
       field='W51 Ku',
@@ -37,7 +38,7 @@ applycal(vis=vis,
          interp='linear',
          flagbackup=True) # was False when flagmanager was used
 delmod(vis=vis)
-pfx = 'h2co_line_35to80kms.crosscal'
+pfx = 'h2co_line_widr.crosscal'
 os.system('rm -rf {0}*'.format(pfx))
 clean(vis=vis, spw='0', imagename=pfx,
       field='W51 Ku',
@@ -47,26 +48,30 @@ clean(vis=vis, spw='0', imagename=pfx,
       selectdata=True)
 exportfits('{0}.image'.format(pfx),'{0}.image.fits'.format(pfx))
 
-pfx = 'h2co_line_35to80kms.crosscal.modelstart'
-os.system('rm -rf {0}*'.format(pfx))
-clean(vis=vis, spw='0', imagename=pfx,
-      field='W51 Ku',
-      modelimage='h2co_cont_4768_to_4880MHz.crosscal.image',
-      weighting='uniform', imsize=[2048,2048], cell=['0.1 arcsec'],
-      mode='channel', threshold='1 mJy', niter=10000,
-      usescratch=True,
-      selectdata=True)
-exportfits('{0}.image'.format(pfx),'{0}.image.fits'.format(pfx))
+if False:
+    # This lead to a segmentation fault: see casapy-20140520-150839.log
+    pfx = 'h2co_line_widr.crosscal.modelstart'
+    os.system('rm -rf {0}*'.format(pfx))
+    clean(vis=vis, spw='0', imagename=pfx,
+          field='W51 Ku',
+          modelimage='h2co_cont_4768_to_4880MHz.crosscal.image',
+          weighting='uniform', imsize=[2048,2048], cell=['0.1 arcsec'],
+          mode='channel', threshold='1 mJy', niter=10000,
+          usescratch=True,
+          selectdata=True)
+    exportfits('{0}.image'.format(pfx),'{0}.image.fits'.format(pfx))
 
 # May 20
-vis = 'h2co_35to80kms'
+vis = 'h2co_widr'
 uvcontsub(vis=vis, field='W51 Ku')
-pfx = 'h2co_line_35to80kms.crosscal.uvcontsub'
+pfx = 'h2co_line_widr.crosscal.uvcontsub'
 os.system('rm -rf {0}*'.format(pfx))
 clean(vis=vis+'.contsub', spw='0', imagename=pfx,
       field='W51 Ku',
       weighting='uniform', imsize=[2048,2048], cell=['0.1 arcsec'],
-      mode='channel', threshold='1 mJy', niter=10000,
+      mode='channel', threshold='0.1 mJy', niter=10000,
+      multiscale=[0,3,6,12,24],
       usescratch=True,
       selectdata=True)
 exportfits('{0}.image'.format(pfx),'{0}.image.fits'.format(pfx))
+
